+++ renderer/units/JavaScript Runner/samples/Unihan/List IDS Indivisible Characters.js
@@ -1,0 +1,18 @@
+// List IDS Indivisible Characters
+const { codePoints } = require ('./lib/unicode/parsed-ids-data.js');
+const { codePointsToCharacters } = require ('./lib/unicode/unicode.js');
+let indivisibleCharacters = [ ];
+for (let codePoint in codePoints)
+{
+    let { sequences } = codePoints[codePoint];
+    let character = codePointsToCharacters (codePoint);
+    for (let sequence of sequences)
+    {
+        if (sequence.ids === character)
+        {
+            indivisibleCharacters.push (character);
+        }
+    }
+}
+$.writeln ("Count:", indivisibleCharacters.length);
+$.writeln (indivisibleCharacters.join (" "));